Release Report System application, all requirements in a task was realized.
New features:
- implemented log process of application events - class Logger. Class Logger
is Singleton with applying RAII programming technics;
- create the ParameterNotFoundBhvr abstract class and its inheritances for
modeling behaviour when any type of parameters from report template was not
registered and/or found. This allow to simplify code and use only one
method (ReportGenerator::findAndInsertParameters()) for finding and
insertion all needed parameters. At now it is required and not required
parameters, but there are an ability to add new behaviours in future
by means of inheritance the ParameterNotFoundBhvr class and overloading
the pure virtual function ParameterNotFoundBhvr::processError();
- show message when report was generated successfully.

Signed-off-by: Oleg <olgerd27@gmail.com>